Implemented traits and debuging
Implemented traits for the AI and debugged attack state transition and movement. Need to get the audio range for AI working and implement audio now

+++ BaseFPCharacter/Assets/Scripts/Sc_QuickSort.cs
@@ -27,7 +27,7 @@ public class Sc_QuickSort : MonoBehaviour
     {
 
         // pivot
-        int pivot = arrayObjects[highEnd].GetComponent<Sc_AIStateManager>().ReturnDecisionValue();
+        float pivot = arrayObjects[highEnd].GetComponent<Sc_AIStateManager>().ReturnDecisionValue();
 
         // Index of smaller element and
         // indicates the right position
